Enhance report and sector handling with new properties
- Added `Original` property to `Report` class and updated its constructor.
- Modified `Sector` and `SectorDetails` constructors to accept new parameters and references.
- Introduced `Version` property in `Template` and `Repeat` classes.
- Updated `ImageResultEntry` and `RunResult` to include `jobVersion`.
- Enhanced `SectorDifferences` to compare `Aperture` and `X Dimension`.
- Changed `Value_` class to use `double` for `Value`.
- Updated XAML bindings for new properties and refactored methods for consistency.

+++ LabelVal/Sectors/Interfaces/ISectorValues.cs
@@ -65,9 +65,9 @@ public class ValueResult : ISectorValue
 public class Value_ : ISectorValue
 {
     public string Name { get; set; }
-    public int Value { get; set; }
+    public double Value { get; set; }
 
-    public Value_(string name, int value)
+    public Value_(string name, double value)
     {
         Value = value;
         Name = name;
